app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Handwritten signature verifier Flask website and API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-wtf>=1.2.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "waitress>=3.0.2",
]
